+++ dev-packages/e2e-tests/test-applications/nuxt-3-top-level-import/playwright.config.ts
@@ -12,7 +12,7 @@ const nuxtConfigOptions: ConfigOptions = {
  *  Like this: import { expect, test } from '@nuxt/test-utils/playwright' */
 
 const config = getPlaywrightConfig({
-  startCommand: `pnpm start:import`,
+  startCommand: `pnpm start`,
   use: { ...nuxtConfigOptions },
 });